refactor files organization
