+++ src/generators/generate-bottom-content.ts
