+++ src/generators/generate-type-from-literal.test.ts
@@ -1,9 +1,9 @@
 import { parse } from 'graphql';
-import { extractTypeFromLiteral } from './extract-type-from-literal';
-import { formatTS } from './test-utils';
+import { generateTypeFromLiteral } from './generate-type-from-literal';
+import { formatTS } from '../utils/test-utils';
 
-describe('Extract type from literal', () => {
-  it('extracts type from correct string', async () => {
+describe('Generate type from literal', () => {
+  it('generates type from correct string', async () => {
     const schema = parse(`
     type User {
       id: ID!
@@ -44,7 +44,7 @@ describe('Extract type from literal', () => {
       type UserQueryOperation = { __typename?: 'Query', user: { __typename?: 'User', id: string, name: string }, users: Array<{ __typename?: 'User', id: string, email: string }> };
     `);
 
-    const result = await extractTypeFromLiteral(code, schema);
+    const result = await generateTypeFromLiteral(code, schema);
 
     expect(result.variables).toEqual(expectedVariables);
     expect(result.result).toEqual(expectedResult);
