+++ src/generators/generate-type-from-literal.ts
@@ -9,7 +9,7 @@ const operationsRegex = /type (\w+Operation) = /s;
 type CodegenPlugin = typeof plugins[number];
 const plugins = [typescriptOperationsPlugin];
 
-export const extractTypeFromLiteral = async (
+export const generateTypeFromLiteral = async (
   literal: string,
   schema: DocumentNode,
   codegenConfig: typescriptOperationsPlugin.TypeScriptDocumentsPluginConfig = {}
